+++ libcaf_core/caf/actor_system_config.cpp
@@ -14,6 +14,7 @@
 #include "caf/detail/parser/read_config.hpp"
 #include "caf/detail/parser/read_string.hpp"
 #include "caf/format_to_error.hpp"
+#include "caf/format_to_unexpected.hpp"
 #include "caf/message_builder.hpp"
 #include "caf/pec.hpp"
 #include "caf/scheduled_actor.hpp"
@@ -541,8 +542,8 @@ actor_system_config::parse_config_file(const char* filename,
                                        const config_option_set& opts) {
   std::ifstream f{filename};
   if (!f.is_open())
-    return format_to_error(sec::cannot_open_file, "cannot open config file: {}",
-                           filename);
+    return format_to_unexpected(sec::cannot_open_file,
+                                "cannot open config file: {}", filename);
   return parse_config(f, opts);
 }
 
@@ -556,7 +557,7 @@ actor_system_config::parse_config(std::istream& source,
                                   const config_option_set& opts) {
   settings result;
   if (auto err = parse_config(source, opts, result); err.valid())
-    return err;
+    return expected<settings>{unexpect, err};
   return result;
 }