+++ libcaf_core/caf/expected.hpp
@@ -9,6 +9,7 @@
 #include "caf/config.hpp"
 #include "caf/deep_to_string.hpp"
 #include "caf/detail/assert.hpp"
+#include "caf/detail/build_config.hpp"
 #include "caf/detail/concepts.hpp"
 #include "caf/error.hpp"
 #include "caf/error_code_enum.hpp"
@@ -22,9 +23,7 @@
 #include <type_traits>
 #include <utility>
 
-namespace caf {
-
-namespace detail {
+namespace caf::detail {
 
 template <class F, class... Ts>
 auto expected_from_fn(F&& f, Ts&&... xs) {
@@ -37,7 +36,82 @@ auto expected_from_fn(F&& f, Ts&&... xs) {
   }
 }
 
-} // namespace detail
+} // namespace caf::detail
+
+#ifdef CAF_USE_STD_EXPECTED
+
+#  include <expected>
+
+namespace caf {
+
+template <class T>
+using unexpected = std::unexpected<T>;
+
+template <class T>
+using expected = std::expected<T, error>;
+
+using unexpect_t = std::unexpect_t;
+
+inline constexpr std::unexpect_t unexpect{};
+
+} // namespace caf
+
+#else
+
+namespace caf {
+
+struct unexpect_t {
+  explicit unexpect_t() = default;
+};
+
+inline constexpr unexpect_t unexpect{};
+
+/// Represents an unexpected value to be stored in caf::expected.
+template <class E>
+class unexpected {
+public:
+  template <typename... Args>
+    requires std::is_constructible_v<error, Args...>
+  constexpr explicit unexpected(std::in_place_t, Args&&... args)
+    : error_{std::forward<Args>(args)...} {
+    // nop
+  }
+
+  template <class U, class... Args>
+    requires std::is_constructible_v<error, std::initializer_list<U>&, Args...>
+  constexpr explicit unexpected(std::in_place_t, std::initializer_list<U> il,
+                                Args&&... args)
+    : error_{std::move(il), std::forward<Args>(args)...} {
+    // nop
+  }
+
+  unexpected(caf::error err) : error_{std::move(err)} {
+  }
+
+  caf::error& error() noexcept {
+    return error_;
+  }
+
+  const caf::error& error() const noexcept {
+    return error_;
+  }
+
+  void swap(unexpected& other) noexcept {
+    using std::swap;
+    swap(error_, other.error_);
+  }
+
+private:
+  E error_;
+};
+
+// TODO requires
+// TODO add comparisons
+template <class E1, class E2>
+inline bool
+operator==(const unexpected<E1>& lhs, const unexpected<E2>& rhs) noexcept {
+  return lhs.error() == rhs.error();
+}
 
 /// Represents the result of a computation which can either complete
 /// successfully with an instance of type `T` or fail with an `error`.
@@ -50,6 +124,8 @@ public:
   using value_type = T;
 
   using error_type = caf::error;
+
+  using unexpected_type = unexpected<caf::error>;
 
   template <class U>
   using rebind = expected<U>;
@@ -70,15 +146,16 @@ public:
   // -- constructors, destructors, and assignment operators --------------------
 
   template <class U>
-    requires(std::convertible_to<U, T> || error_code_enum<U>)
-  expected(U x) {
-    if constexpr (std::convertible_to<U, T>) {
-      has_value_ = true;
-      new (std::addressof(value_)) T(std::move(x));
-    } else {
-      has_value_ = false;
-      new (std::addressof(error_)) caf::error(x);
-    }
+    requires std::is_constructible_v<T, U>
+  expected(U x) : has_value_{true} {
+    new (std::addressof(value_)) T(std::move(x));
+  }
+
+  template <error_code_enum U>
+  [[deprecated("construct using unexpect or from an unexpected instead")]]
+  expected(U x)
+    : has_value_{false} {
+    new (std::addressof(error_)) caf::error(x);
   }
 
   expected(T&& x) noexcept(nothrow_move) : has_value_(true) {
@@ -89,8 +166,14 @@ public:
     new (std::addressof(value_)) T(x);
   }
 
-  expected(caf::error e) noexcept : has_value_(false) {
+  [[deprecated("construct using unexpect or from an unexpected instead")]]
+  expected(caf::error e) noexcept
+    : has_value_(false) {
     new (std::addressof(error_)) caf::error{std::move(e)};
+  }
+
+  expected(unexpected_type x) : has_value_(false) {
+    new (std::addressof(error_)) caf::error(std::move(x.error()));
   }
 
   expected(const expected& other) noexcept(nothrow_copy) {
@@ -104,6 +187,11 @@ public:
   template <class... Ts>
   expected(std::in_place_t, Ts&&... xs) : has_value_(true) {
     new (std::addressof(value_)) T(std::forward<Ts>(xs)...);
+  }
+
+  template <class... Ts>
+  explicit expected(caf::unexpect_t, Ts&&... xs) : has_value_(false) {
+    new (std::addressof(error_)) caf::error(std::forward<Ts>(xs)...);
   }
 
   ~expected() {
@@ -156,11 +244,13 @@ public:
     return *this;
   }
 
-  template <std::convertible_to<T> U>
+  template <class U>
+    requires std::is_constructible_v<T, U>
   expected& operator=(U x) {
     return *this = T{std::move(x)};
   }
 
+  [[deprecated("use assignment with caf::unexpected instead")]]
   expected& operator=(caf::error e) noexcept {
     if (!has_value())
       error_ = std::move(e);
@@ -173,8 +263,20 @@ public:
   }
 
   template <error_code_enum Enum>
+  [[deprecated("use assignment with caf::unexpected instead")]]
   expected& operator=(Enum code) {
     return *this = make_error(code);
+  }
+
+  expected& operator=(unexpected_type e) noexcept {
+    if (!has_value())
+      error_ = std::move(e.error());
+    else {
+      destroy();
+      has_value_ = false;
+      new (std::addressof(error_)) caf::error(std::move(e.error()));
+    }
+    return *this;
   }
 
   // -- observers --------------------------------------------------------------
@@ -321,7 +423,7 @@ public:
     if (has_value())
       return f(value_);
     else
-      return res_t{error_};
+      return res_t{unexpect, error_};
   }
 
   template <class F>
@@ -331,7 +433,7 @@ public:
     if (has_value())
       return f(std::move(value_));
     else
-      return res_t{std::move(error_)};
+      return res_t{unexpect, std::move(error_)};
   }
 
   template <class F>
@@ -341,7 +443,7 @@ public:
     if (has_value())
       return f(value_);
     else
-      return res_t{error_};
+      return res_t{unexpect, error_};
   }
 
   template <class F>
@@ -351,75 +453,87 @@ public:
     if (has_value())
       return f(std::move(value_));
     else
-      return res_t{std::move(error_)};
+      return res_t{unexpect, std::move(error_)};
   }
 
   template <class F>
+    requires std::is_void_v<
+      decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  [[deprecated("use or_else with expected return type instead")]]
   expected or_else(F&& f) & {
-    using res_t = decltype(f(error_));
-    if constexpr (std::is_void_v<res_t>) {
-      if (!has_value())
-        f(error_);
-      return *this;
-    } else {
-      static_assert(std::is_same_v<expected, res_t>,
-                    "F must return expected<T> or void");
-      if (has_value())
-        return expected{std::in_place, value_};
-      else
-        return f(error_);
-    }
+    if (!has_value())
+      f(error_);
+    return *this;
   }
 
   template <class F>
+    requires std::is_same_v<
+      expected, decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  expected or_else(F&& f) & {
+    if (has_value())
+      return expected{std::in_place, value_};
+    else
+      return f(error_);
+  }
+
+  template <class F>
+    requires std::is_void_v<
+      decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  [[deprecated("use or_else with expected return type instead")]]
   expected or_else(F&& f) && {
-    using res_t = decltype(f(std::move(error_)));
-    if constexpr (std::is_void_v<res_t>) {
-      if (!has_value())
-        f(std::move(error_));
-      return std::move(*this);
-    } else {
-      static_assert(std::is_same_v<expected, res_t>,
-                    "F must return expected<T> or void");
-      if (has_value())
-        return expected{std::in_place, std::move(value_)};
-      else
-        return f(std::move(error_));
-    }
+    if (!has_value())
+      f(std::move(error_));
+    return std::move(*this);
   }
 
   template <class F>
+    requires std::is_same_v<
+      expected, decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  expected or_else(F&& f) && {
+    if (has_value())
+      return expected{std::in_place, std::move(value_)};
+    else
+      return f(std::move(error_));
+  }
+
+  template <class F>
+    requires std::is_void_v<
+      decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  [[deprecated("use or_else with expected return type instead")]]
   expected or_else(F&& f) const& {
-    using res_t = decltype(f(error_));
-    if constexpr (std::is_void_v<res_t>) {
-      if (!has_value())
-        f(error_);
-      return *this;
-    } else {
-      static_assert(std::is_same_v<expected, res_t>,
-                    "F must return expected<T> or void");
-      if (has_value())
-        return expected{std::in_place, value_};
-      else
-        return f(error_);
-    }
+    if (!has_value())
+      f(error_);
+    return *this;
   }
 
   template <class F>
+    requires std::is_same_v<
+      expected, decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  expected or_else(F&& f) const& {
+    if (has_value())
+      return expected{std::in_place, value_};
+    else
+      return f(error_);
+  }
+
+  template <class F>
+    requires std::is_void_v<
+      decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  [[deprecated("use or_else with expected return type instead")]]
   expected or_else(F&& f) const&& {
-    using res_t = decltype(f(std::move(error_)));
-    if constexpr (std::is_void_v<res_t>) {
-      if (!has_value())
-        f(std::move(error_));
-      return std::move(*this);
-    } else {
-      static_assert(std::is_same_v<expected, res_t>,
-                    "F must return expected<T> or void");
-      if (has_value())
-        return expected{std::in_place, std::move(value_)};
-      else
-        return f(std::move(error_));
-    }
+    if (!has_value())
+      f(std::move(error_));
+    return std::move(*this);
+  }
+
+  template <class F>
+    requires std::is_same_v<
+      expected, decltype(std::declval<F&>()(std::declval<caf::error>()))>
+  expected or_else(F&& f) const&& {
+    if (has_value())
+      return expected{std::in_place, std::move(value_)};
+    else
+      return f(std::move(error_));
   }
 
   template <class F>
@@ -429,7 +543,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f), value_);
     else
-      return expected<res_t>{error_};
+      return expected<res_t>{unexpect, error_};
   }
 
   template <class F>
@@ -439,7 +553,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f), std::move(value_));
     else
-      return expected<res_t>{std::move(error_)};
+      return expected<res_t>{unexpect, std::move(error_)};
   }
 
   template <class F>
@@ -449,7 +563,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f), value_);
     else
-      return expected<res_t>{error_};
+      return expected<res_t>{unexpect, error_};
   }
 
   template <class F>
@@ -459,47 +573,71 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f), std::move(value_));
     else
-      return expected<res_t>{std::move(error_)};
+      return expected<res_t>{unexpect, std::move(error_)};
   }
 
   template <class F>
+  [[deprecated("use transform_error")]]
   expected transform_or(F&& f) & {
-    using res_t = decltype(f(error_));
-    static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
-    if (has_value())
-      return {std::in_place, value_};
-    else
-      return expected{f(error_)};
+    return transform_error(std::forward<F>(f));
   }
 
   template <class F>
+  [[deprecated("use transform_error")]]
   expected transform_or(F&& f) && {
-    using res_t = decltype(f(error_));
-    static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
-    if (has_value())
-      return {std::in_place, std::move(value_)};
-    else
-      return expected{f(std::move(error_))};
+    return std::move(*this).transform_error(std::forward<F>(f));
   }
 
   template <class F>
+  [[deprecated("use transform_error")]]
   expected transform_or(F&& f) const& {
+    return transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  [[deprecated("use transform_error")]]
+  expected transform_or(F&& f) const&& {
+    return std::move(*this).transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  expected transform_error(F&& f) & {
     using res_t = decltype(f(error_));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return {std::in_place, value_};
     else
-      return expected{f(error_)};
+      return expected{unexpect, f(error_)};
   }
 
   template <class F>
-  expected transform_or(F&& f) const&& {
+  expected transform_error(F&& f) && {
     using res_t = decltype(f(error_));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return {std::in_place, std::move(value_)};
     else
-      return expected{f(std::move(error_))};
+      return expected{unexpect, f(std::move(error_))};
+  }
+
+  template <class F>
+  expected transform_error(F&& f) const& {
+    using res_t = decltype(f(error_));
+    static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
+    if (has_value())
+      return {std::in_place, value_};
+    else
+      return expected{unexpect, f(error_)};
+  }
+
+  template <class F>
+  expected transform_error(F&& f) const&& {
+    using res_t = decltype(f(error_));
+    static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
+    if (has_value())
+      return {std::in_place, std::move(value_)};
+    else
+      return expected{unexpect, f(std::move(error_))};
   }
 
 private:
@@ -560,26 +698,43 @@ auto operator==(const T& x, const expected<U>& y) -> decltype(x == *y) {
 
 /// @relates expected
 template <class T>
+[[deprecated("Compare with caf::unexpected{error} instead")]]
 bool operator==(const expected<T>& x, const error& y) {
   return x ? false : x.error() == y;
 }
 
 /// @relates expected
 template <class T>
+[[deprecated("Compare with caf::unexpected{error} instead")]]
 bool operator==(const error& x, const expected<T>& y) {
   return y == x;
 }
 
 /// @relates expected
 template <class T, error_code_enum Enum>
+[[deprecated("Compare with caf::unexpected{error{enum}} instead")]]
 bool operator==(const expected<T>& x, Enum y) {
   return x == make_error(y);
 }
 
 /// @relates expected
 template <error_code_enum Enum, class T>
+[[deprecated("Compare with caf::unexpected{error{enum}} instead")]]
 bool operator==(Enum x, const expected<T>& y) {
   return y == make_error(x);
+}
+
+template <class T, class E>
+// TODO requires
+bool operator==(const expected<T>& x, const unexpected<E>& y) {
+  return x ? false : x.error() == y.error();
+}
+
+/// @relates expected
+template <class E, class T>
+// TODO requires
+bool operator==(const unexpected<E>& x, const expected<T>& y) {
+  return y == x;
 }
 
 /// @relates expected
@@ -603,25 +758,43 @@ auto operator!=(const T& x, const expected<U>& y) -> decltype(x == *y) {
 
 /// @relates expected
 template <class T>
+[[deprecated("Compare with caf::unexpected{error} instead")]]
 bool operator!=(const expected<T>& x, const error& y) {
   return !(x == y);
 }
 
 /// @relates expected
 template <class T>
+[[deprecated("Compare with caf::unexpected{error} instead")]]
 bool operator!=(const error& x, const expected<T>& y) {
   return !(x == y);
 }
 
 /// @relates expected
 template <class T, error_code_enum Enum>
+[[deprecated("Compare with caf::unexpected{error{enum}} instead")]]
 bool operator!=(const expected<T>& x, Enum y) {
   return !(x == y);
 }
 
 /// @relates expected
 template <error_code_enum Enum, class T>
+[[deprecated("Compare with caf::unexpected{error{enum}} instead")]]
 bool operator!=(Enum x, const expected<T>& y) {
+  return !(x == y);
+}
+
+/// @relates expected
+template <class T, class E>
+// TODO requires
+bool operator!=(const expected<T>& x, const unexpected<E>& y) {
+  return !(x == y);
+}
+
+/// @relates expected
+template <class E, class T>
+// TODO
+bool operator!=(const unexpected<E>& x, const expected<T>& y) {
   return !(x == y);
 }
 
@@ -636,19 +809,29 @@ public:
 
   using error_type = caf::error;
 
+  using unexpected_type = unexpected<caf::error>;
+
   template <class U>
   using rebind = expected<U>;
 
   // -- constructors, destructors, and assignment operators --------------------
 
   template <error_code_enum Enum>
-  expected(Enum x) : error_(std::in_place, x) {
+  [[deprecated("construct using unexpect or from an unexpected instead")]]
+  expected(Enum x)
+    : error_(std::in_place, x) {
+    // nop
+  }
+
+  [[deprecated("construct using unexpect or from an unexpected instead")]]
+  expected(caf::error err) noexcept
+    : error_(std::move(err)) {
     // nop
   }
 
   expected() noexcept = default;
 
-  expected(caf::error err) noexcept : error_(std::move(err)) {
+  expected(unexpected_type x) noexcept : error_(std::move(x.error())) {
     // nop
   }
 
@@ -660,18 +843,31 @@ public:
     // nop
   }
 
+  template <class... Args>
+  explicit expected(unexpect_t, Args&&... args) noexcept
+    : error_(std::forward<Args>(args)...) {
+    // nop
+  }
+
   expected& operator=(const expected& other) = default;
 
   expected& operator=(expected&& other) noexcept = default;
 
+  [[deprecated("use assignment with caf::unexpected instead")]]
   expected& operator=(caf::error err) noexcept {
     error_ = std::move(err);
     return *this;
   }
 
   template <error_code_enum Enum>
+  [[deprecated("use assignment with caf::unexpected instead")]]
   expected& operator=(Enum code) {
     error_ = make_error(code);
+    return *this;
+  }
+
+  expected& operator=(unexpected_type x) {
+    error_ = std::move(x.error());
     return *this;
   }
 
@@ -743,7 +939,7 @@ public:
     if (has_value())
       return f();
     else
-      return res_t{*error_};
+      return res_t{unexpect, *error_};
   }
 
   template <class F>
@@ -753,7 +949,7 @@ public:
     if (has_value())
       return f();
     else
-      return res_t{std::move(*error_)};
+      return res_t{unexpect, std::move(*error_)};
   }
 
   template <class F>
@@ -763,7 +959,7 @@ public:
     if (has_value())
       return f();
     else
-      return res_t{*error_};
+      return res_t{unexpect, *error_};
   }
 
   template <class F>
@@ -773,7 +969,7 @@ public:
     if (has_value())
       return f();
     else
-      return res_t{*error_};
+      return res_t{unexpect, *error_};
   }
 
   template <class F>
@@ -851,7 +1047,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f));
     else
-      return expected<res_t>{*error_};
+      return expected<res_t>{unexpect, *error_};
   }
 
   template <class F>
@@ -861,7 +1057,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f));
     else
-      return expected<res_t>{std::move(*error_)};
+      return expected<res_t>{unexpect, std::move(*error_)};
   }
 
   template <class F>
@@ -871,7 +1067,7 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f));
     else
-      return expected<res_t>{*error_};
+      return expected<res_t>{unexpect, *error_};
   }
 
   template <class F>
@@ -881,47 +1077,71 @@ public:
     if (has_value())
       return detail::expected_from_fn(std::forward<F>(f));
     else
-      return expected<res_t>{*error_};
+      return expected<res_t>{unexpect, *error_};
   }
 
   template <class F>
+  [[deprecated("use transform_error")]]
   expected transform_or(F&& f) & {
+    return transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  [[deprecated("use transform_error")]]
+  expected transform_or(F&& f) && {
+    return std::move(*this).transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  [[deprecated("use transform_error")]]
+  expected transform_or(F&& f) const& {
+    return transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  [[deprecated("use transform_error")]]
+  expected transform_or(F&& f) const&& {
+    return std::move(*this).transform_error(std::forward<F>(f));
+  }
+
+  template <class F>
+  expected transform_error(F&& f) & {
     using res_t = decltype(f(*error_));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return expected{};
     else
-      return expected{f(*error_)};
+      return expected{unexpect, f(*error_)};
   }
 
   template <class F>
-  expected transform_or(F&& f) && {
+  expected transform_error(F&& f) && {
     using res_t = decltype(f(std::move(*error_)));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return expected{};
     else
-      return expected{f(std::move(*error_))};
+      return expected{unexpect, f(std::move(*error_))};
   }
 
   template <class F>
-  expected transform_or(F&& f) const& {
+  expected transform_error(F&& f) const& {
     using res_t = decltype(f(*error_));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return expected{std::in_place};
     else
-      return expected{f(*error_)};
+      return expected{unexpect, f(*error_)};
   }
 
   template <class F>
-  expected transform_or(F&& f) const&& {
+  expected transform_error(F&& f) const&& {
     using res_t = decltype(f(std::move(*error_)));
     static_assert(std::is_same_v<caf::error, res_t>, "F must return an error");
     if (has_value())
       return expected{};
     else
-      return expected{f(std::move(*error_))};
+      return expected{unexpect, f(std::move(*error_))};
   }
 
 private:
@@ -936,6 +1156,22 @@ inline bool operator==(const expected<void>& x, const expected<void>& y) {
 /// @relates expected
 inline bool operator!=(const expected<void>& x, const expected<void>& y) {
   return !(x == y);
+}
+
+} // namespace caf
+
+#endif
+
+namespace caf {
+
+template <error_code_enum Enum, class... Args>
+unexpected<error> make_unexpected(Enum e, Args&&... args) noexcept {
+  return unexpected<error>{std::in_place,
+                           make_error(e, std::forward<Args>(args)...)};
+}
+
+inline unexpected<error> make_unexpected(error err) noexcept {
+  return unexpected<error>{std::move(err)};
 }
 
 template <class T>