+++ libcaf_core/caf/settings.hpp
@@ -85,7 +85,7 @@ expected<T> get_as(const settings& xs, std::string_view name) {
   if (auto ptr = get_if(&xs, name))
     return get_as<T>(*ptr);
   else
-    return {sec::no_such_key};
+    return expected<T>{unexpect, sec::no_such_key};
 }
 
 /// @private