+++ libcaf_core/caf/uuid.cpp
@@ -223,7 +223,7 @@ std::string to_string(const uuid& x) {
 expected<uuid> make_uuid(std::string_view str) {
   uuid result;
   if (auto err = parse(str, result); err.valid())
-    return err;
+    return expected<uuid>{unexpect, err};
   return result;
 }