+++ libcaf_io/caf/detail/call_cfun.hpp
@@ -8,6 +8,7 @@
 
 #include "caf/error.hpp"
 #include "caf/format_to_error.hpp"
+#include "caf/format_to_unexpected.hpp"
 #include "caf/sec.hpp"
 
 #include <cstdio>
@@ -39,8 +40,8 @@ inline bool cc_valid_socket(caf::io::network::native_socket fd) {
 #define CALL_CFUN(var, predicate, fun_name, expr)                              \
   auto var = expr;                                                             \
   if (!predicate(var))                                                         \
-  return format_to_error(sec::network_syscall_failed, "{}: {}", fun_name,      \
-                         last_socket_error_as_string())
+  return format_to_unexpected(sec::network_syscall_failed, "{}: {}", fun_name, \
+                              last_socket_error_as_string())
 
 /// Calls a C functions and calls exit() if `predicate(var)` returns false.
 #define CALL_CRITICAL_CFUN(var, predicate, funname, expr)                      \