+++ libcaf_net/caf/detail/ws_conn_acceptor.hpp
@@ -72,8 +72,9 @@ public:
 
   expected<resources> start() override {
     if (!producer_->push(event_)) {
-      return make_error(sec::runtime_error,
-                        "WebSocket connection dropped: client canceled");
+      return expected<resources>{
+        unexpect, sec::runtime_error,
+        "WebSocket connection dropped: client canceled"};
     }
     return res_;
   }
@@ -104,8 +105,9 @@ public:
   expected<ws_conn_starter_ptr> accept(const net::http::request_header& hdr,
                                        net::socket_manager* mgr) override {
     if (!producer_) {
-      return make_error(sec::runtime_error,
-                        "WebSocket connection dropped: client canceled");
+      return expected<ws_conn_starter_ptr>{
+        unexpect, sec::runtime_error,
+        "WebSocket connection dropped: client canceled"};
     }
     ws_acceptor_impl<Ts...> acc{hdr, mgr};
     on_request_(acc);
@@ -116,10 +118,11 @@ public:
       return res;
     }
     if (auto&& reason = std::move(acc).reject_reason(); reason.valid()) {
-      return std::move(reason);
+      return expected<ws_conn_starter_ptr>{unexpect, std::move(reason)};
     }
-    return make_error(sec::runtime_error,
-                      "WebSocket request rejected without reason");
+    return expected<ws_conn_starter_ptr>{
+      unexpect, sec::runtime_error,
+      "WebSocket request rejected without reason"};
   }
 
   bool canceled() const noexcept override {