+++ libcaf_net/caf/internal/net_syscall.hpp
@@ -6,6 +6,7 @@
 
 #include "caf/error.hpp"
 #include "caf/format_to_error.hpp"
+#include "caf/format_to_unexpected.hpp"
 #include "caf/sec.hpp"
 
 #include <cstdio>
@@ -18,6 +19,15 @@
   return format_to_error(sec::network_syscall_failed,                          \
                          "error in function {}: {}", funname,                  \
                          last_socket_error_as_string())
+
+/// Calls a C functions and returns an caf::unexpected if `var op rhs` returns
+/// `true`.
+#define CAF_NET_SYSCALL_TO_UNEXPECTED(funname, var, op, rhs, expr)             \
+  auto var = expr;                                                             \
+  if (var op rhs)                                                              \
+  return format_to_unexpected(sec::network_syscall_failed,                     \
+                              "error in function {}: {}", funname,             \
+                              last_socket_error_as_string())
 
 /// Calls a C functions and calls exit() if `var op rhs` returns `true`.
 #define CAF_NET_CRITICAL_SYSCALL(funname, var, op, rhs, expr)                  \